Pinterest_Scraper/requirements.txt
aiohttp==3.8.1
beautifulsoup4==4.11.1
numpy==1.22.2
pandas==1.4.1
python-dotenv==0.20.0
pyinstaller==5.0.1
